Add VC++ compiler support.

+++ src/utils/utility.cpp
@@ -4,7 +4,7 @@
 
 uint strToByte_hex(const string& input, ubyte* result, uint max_len)
 {
-    uint len = (max_len <= 0 ? input.length() : max_len) / 2;
+    uint len = (uint)((max_len <= 0 ? input.length() : max_len) / 2);
     char* p = (char*)input.c_str();
 
     for (uint i = 0 ; i < len; i++)
